refactor(schema): move actions payload types to @contember/schema
